Added separate field for completing onboarding vs call
