+++ prisma/schema.prisma
@@ -12,6 +12,7 @@ model User {
     email         String    @unique
     emailVerified DateTime?
     onboarded     Boolean?  @default(false)
+    verified      Boolean?  @default(false)
     name          String?   @default("")
     bio           String?   @default("")
     image         String    @default("")
