+++ src/server/api/routers/user.ts
@@ -10,30 +10,33 @@ import { type OnboardStepOneValues } from "@/lib/types";
 export const userRoute = createTRPCRouter({
   updateUser: protectedProcedure
     .input(
-      z.object({
-        name: z
-          .string()
-          .min(2, "Must be at least 2 characters.")
-          .max(128, "Can not exceed 128 characters."),
-        phone: z
-          .string({
-            required_error: "Can not be empty.",
-            invalid_type_error: "Can not be empty.",
-          })
-          .refine(
-            (phone) => isValidPhoneNumber(phone),
-            "Must be a valid phone number.",
-          ),
-        graduation: z
-          .string()
-          .regex(/^(0[1-9]|1[0-2])\/([0-9]{2})$/g, "Must be a valid date."),
-        major: z
-          .string()
-          .min(4, "Must be at least 4 characters.")
-          .max(128, "Can not exceed 32 characters.")
-          .regex(/^[a-zA-Z- ]+$/, "Must be a valid major"),
-        idea: z.string().max(128, "Can not exceed 128 characters."),
-      }),
+      z
+        .object({
+          name: z
+            .string()
+            .min(2, "Must be at least 2 characters.")
+            .max(128, "Can not exceed 128 characters."),
+          phone: z
+            .string({
+              required_error: "Can not be empty.",
+              invalid_type_error: "Can not be empty.",
+            })
+            .refine(
+              (phone) => isValidPhoneNumber(phone),
+              "Must be a valid phone number.",
+            ),
+          graduation: z
+            .string()
+            .regex(/^(0[1-9]|1[0-2])\/([0-9]{2})$/g, "Must be a valid date."),
+          major: z
+            .string()
+            .min(4, "Must be at least 4 characters.")
+            .max(128, "Can not exceed 32 characters.")
+            .regex(/^[a-zA-Z- ]+$/, "Must be a valid major"),
+          idea: z.string().max(128, "Can not exceed 128 characters."),
+          onboarded: z.boolean(),
+        })
+        .partial(),
     )
     .mutation(async ({ input, ctx }) => {
       try {
